Cast type explicitly for fix build error (#111)

+++ src/backward.h
@@ -3266,7 +3266,7 @@ public:
         if (st.size() == 0) {
             return;
         }
-        _symbols.reset(backtrace_symbols(st.begin(), st.size()));
+        _symbols.reset(backtrace_symbols(st.begin(), (int)st.size()));
     }
 
     ResolvedTrace resolve(ResolvedTrace trace)
